CustomerLogic Tests

+++ quali-t-app/app/logics/project/Customer.java
@@ -1,6 +1,5 @@
 package logics.project;
 
-import com.fasterxml.jackson.databind.JsonNode;
 import dao.models.CustomerDAO;
 import exceptions.EntityAlreadyExistsException;
 import exceptions.EntityNotFoundException;
@@ -17,16 +16,31 @@ public class Customer {
     static CustomerDAO customerDAO = new CustomerDAO();
 
     public static boolean validate(String name) {
-        return !name.equals("");
+        return name != null && !name.isEmpty();
     }
 
-    public static models.project.Customer createCustomer(JsonNode json) throws EntityAlreadyExistsException, MissingParameterException {
-        models.project.Customer customer = new models.project.Customer(json.findPath("name").asText(), json.findPath("address").asText());
-        if (validate(customer.getName())) {
+    public static models.project.Customer createCustomer(models.project.Customer customer) throws EntityAlreadyExistsException, MissingParameterException {
+        if (customer != null && validate(customer.getName())) {
             models.project.Customer c = customerDAO.findByName(customer.getName());
             if (c == null) {
-                customer.setProjects(null);
+                customer.setId(null);
                 return customerDAO.persist(customer);
+            } else {
+                throw new EntityAlreadyExistsException("Customer name already exists");
+            }
+            } else {
+            throw new MissingParameterException("Missing required paramter");
+            }
+    }
+
+    public static models.project.Customer updateCustomer(models.project.Customer updatedCustomer) throws MissingParameterException, EntityNotFoundException, EntityAlreadyExistsException {
+        if (updatedCustomer != null && updatedCustomer.getId() != null && validate(updatedCustomer.getName())) {
+            models.project.Customer c = customerDAO.findByName(updatedCustomer.getName());
+            if (c == null) {
+                models.project.Customer customer = customerDAO.readById(updatedCustomer.getId());
+                customer.setAddress(updatedCustomer.getAddress());
+                customer.setName(updatedCustomer.getName());
+                return customerDAO.update(customer);
             } else {
                 throw new EntityAlreadyExistsException("Customer name already exists");
             }
@@ -35,26 +49,19 @@ public class Customer {
         }
     }
 
-    public static models.project.Customer updateCustomer(JsonNode json) throws MissingParameterException, EntityNotFoundException {
-        models.project.Customer customer = customerDAO.readById(json.findPath("id").asLong());
-        customer.setAddress(json.findPath("address").asText());
-        customer.setName(json.findPath("name").asText());
-        if (validate(customer.getName())) {
-            return customerDAO.update(customer);
-        } else {
-            throw new MissingParameterException("Name can not be empty");
-        }
-    }
-
     public static List<models.project.Customer> getAllCustomers() {
         return customerDAO.readAll();
     }
 
-    public static void deleteCustomer(Long id) throws EntityNotFoundException {
-        models.project.Customer customer = customerDAO.readById(id);
-        for (models.project.Project project : customer.getProjects()) {
-            deleteProject(project.getId());
+    public static void deleteCustomer(Long id) throws EntityNotFoundException, MissingParameterException {
+        if (id != null) {
+            models.project.Customer customer = customerDAO.readById(id);
+            for (models.project.Project project : customer.getProjects()) {
+                deleteProject(project.getId());
+            }
+            customerDAO.remove(customer);
+        } else {
+            throw new MissingParameterException("ID can't be null!");
         }
-        customerDAO.remove(customer);
     }
 }
